+++ CountlyConnectionManager.h
@@ -21,6 +21,7 @@ extern NSString* const kCountlyEndpointFeedback;
 extern NSString* const kCountlyEndpointWidget;
 extern NSString* const kCountlyEndpointSurveys;
 extern NSString* const kCountlyRCKeyKeys;
+extern NSString* const kCountlyQSKeyTimestamp;
 
 extern const NSInteger kCountlyGETRequestMaxLength;